splitting tag models and converting es6
