+++ public/models/entry.js
